+++ android/versioned-abis/expoview-abi38_0_0/src/main/java/abi38_0_0/org/unimodules/interfaces/facedetector/FaceDetectionError.java
@@ -1,0 +1,7 @@
+package abi38_0_0.org.unimodules.interfaces.facedetector;
+
+public interface FaceDetectionError {
+
+  void onError(Throwable error);
+
+}