+++ apps/native-component-list/src/screens/Image/CompareBar.tsx
@@ -1,23 +1,31 @@
 import { MaterialIcons } from '@expo/vector-icons';
 import * as React from 'react';
-import { StyleSheet, TouchableOpacity, Text } from 'react-native';
+import { StyleSheet, TouchableOpacity, Text, View } from 'react-native';
 
 import Colors from '../../constants/Colors';
 
 type PropsType = {
   collapsed: boolean;
+  ImageComponent: React.ComponentType<any>;
   onPress: (collapsed: boolean) => any;
+  onPressComponent: (ImageComponent: React.ComponentType<any>) => any;
 };
 
 export default class CompareBar extends React.Component<PropsType> {
   render() {
-    const { collapsed } = this.props;
+    const { collapsed, ImageComponent } = this.props;
+    const name = ImageComponent.displayName || ImageComponent.name || 'Component';
     return (
       <TouchableOpacity
         style={[styles.container, collapsed ? styles.collapsed : undefined]}
         activeOpacity={0.5}
         onPress={this.onPress}>
-        <Text style={styles.heading}>{'Compare to <Image>'}</Text>
+        <View style={styles.row}>
+          <Text style={styles.heading}>Compare to:</Text>
+          <TouchableOpacity onPress={this.onPressComponent}>
+            <Text style={styles.component}>{`<${name}>`}</Text>
+          </TouchableOpacity>
+        </View>
         <MaterialIcons name={collapsed ? 'keyboard-arrow-down' : 'keyboard-arrow-up'} size={25} />
       </TouchableOpacity>
     );
@@ -27,12 +35,16 @@ export default class CompareBar extends React.Component<PropsType> {
     const { collapsed, onPress } = this.props;
     if (onPress) onPress(!collapsed);
   };
+
+  onPressComponent = () => {
+    const { ImageComponent, onPressComponent } = this.props;
+    if (onPressComponent) onPressComponent(ImageComponent);
+  };
 }
 
 const styles = StyleSheet.create({
   container: {
     flexDirection: 'row',
-    justifyContent: 'space-between',
     alignItems: 'center',
     paddingHorizontal: 16,
     height: 44,
@@ -41,11 +53,23 @@ const styles = StyleSheet.create({
     borderBottomColor: Colors.border,
     borderBottomWidth: StyleSheet.hairlineWidth,
   },
-  collapsed: {
-    borderBottomWidth: 0,
+  row: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
   },
   heading: {
     fontWeight: 'bold',
     fontSize: 18,
+  },
+  component: {
+    color: Colors.highlightColor,
+    fontWeight: 'bold',
+    fontSize: 18,
+    marginHorizontal: 10,
+    marginVertical: 4,
+  },
+  collapsed: {
+    borderBottomWidth: 0,
   },
 });