+++ docs/static/examples/v37.0.0/tutorial/sharing-simple.js
@@ -28,7 +28,7 @@ export default function App() {
       return;
     }
 
-    Sharing.shareAsync(selectedImage.localUri);
+    await Sharing.shareAsync(selectedImage.localUri);
   };
 
   if (selectedImage !== null) {