+++ ios/Pods/Headers/Private/ABI38_0_0ExpoKit/ABI38_0_0RNPinchHandler.h
@@ -1,0 +1,1 @@
+../../../../versioned-react-native/ABI38_0_0/Expo/ExpoKit/Core/Api/Components/GestureHandler/Handlers/ABI38_0_0RNPinchHandler.h