+++ ios/Pods/Headers/Public/ABI38_0_0React-Core/ABI38_0_0React/ABI38_0_0RCTRawTextShadowView.h
@@ -1,0 +1,1 @@
+../../../../../versioned-react-native/ABI38_0_0/ReactNative/Libraries/Text/RawText/ABI38_0_0RCTRawTextShadowView.h