+++ ios/Pods/Target Support Files/ABI38_0_0UMAppLoader/ABI38_0_0UMAppLoader-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_ABI38_0_0UMAppLoader : NSObject
+@end
+@implementation PodsDummy_ABI38_0_0UMAppLoader
+@end