+++ packages/eslint-config-universe/shared/react.js
@@ -36,7 +36,6 @@ module.exports = {
     'react/react-in-jsx-scope': 'error',
     'react/require-render-return': 'warn',
     'react/self-closing-comp': 'warn',
-    'react/style-prop-object': 'warn',
 
     'react-hooks/rules-of-hooks': 'error',
     'react-hooks/exhaustive-deps': 'warn',