+++ packages/expo-auth-session/src/AuthSession.ts
@@ -24,6 +24,7 @@ import {
   ProviderMetadata,
   resolveDiscoveryAsync,
 } from './Discovery';
+import { generateHexStringAsync } from './PKCE';
 import { getQueryParams } from './QueryParams';
 import { getSessionUrlProvider } from './SessionUrlProvider';
 
@@ -203,4 +204,5 @@ export {
   ResponseType,
   resolveDiscoveryAsync,
   fetchDiscoveryAsync,
+  generateHexStringAsync,
 };