+++ packages/expo-blur/src/BlurView.android.tsx
@@ -1,27 +1,14 @@
-import PropTypes from 'prop-types';
 import * as React from 'react';
-import { View, ViewPropTypes } from 'react-native';
+import { View } from 'react-native';
 
-import { BlurTint, BlurProps } from './BlurView.types';
+import { BlurProps } from './BlurView.types';
 import getBackgroundColor from './getBackgroundColor';
 
-export default class BlurView extends React.Component<BlurProps> {
-  static propTypes = {
-    ...ViewPropTypes,
-    tint: PropTypes.oneOf(['light', 'default', 'dark'] as BlurTint[]).isRequired,
-    intensity: PropTypes.number.isRequired,
-  };
-
-  static defaultProps = {
-    tint: 'default' as BlurTint,
-    intensity: 100,
-  };
-
-  render() {
-    const { tint, intensity, ...props } = this.props;
-
+const BlurView = React.forwardRef<View, BlurProps>(
+  ({ tint = 'default', intensity = 100, style, ...props }, ref) => {
     const backgroundColor = getBackgroundColor(intensity, tint);
-
-    return <View {...props} style={[this.props.style, { backgroundColor }]} />;
+    return <View {...props} ref={ref} style={[style, { backgroundColor }]} />;
   }
-}
+);
+
+export default BlurView;