+++ packages/expo-updates/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### 🐛 Bug fixes
 
+- Fixed issue where launch screen on iOS doesn't show whilst updates are being retrieved if it is contained within a storyboard instead of a nib.
+
 ## 0.2.8 — 2020-05-29
 
 *This version does not introduce any user-facing changes.*